refactor: svg 경로 절대 경로로 변경
